% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking-estimation.R
\name{compute_model_weights}
\alias{compute_model_weights}
\title{Compute (log) weights of component models based on an xgbstack fit and
new data.}
\usage{
compute_model_weights(xgbstack_fit, newdata, ntreelimit, log = FALSE,
  format = "bare")
}
\arguments{
\item{xgbstack_fit}{a fit xgbstack object}

\item{newdata}{new x data}

\item{ntreelimit}{how many boosting iterations worth of trees to use}

\item{log}{boolean: return log of weights or original weights?}

\item{format}{string, either "bare" to return an nrow(newdata) by num_models
matrix of (log) weights, or "complete" to return a data frame with
nrow(newdata) times num_models data frame with both weights and log_weights
as well as parameter values and newdata}
}
\value{
(log) weights in the format determined by format
}
\description{
Format "complete" is only partially implemented.
}

