% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking-estimation.R
\name{compute_model_weights_from_preds}
\alias{compute_model_weights_from_preds}
\title{Function to compute (log) weights of component models from the "predictions"
output by xgboost}
\usage{
compute_model_weights_from_preds(preds, log = FALSE)
}
\arguments{
\item{preds}{predictions from xgboost in matrix form, with num_models columns
and num_obs rows}

\item{log}{boolean; return log of component model weights or just the weights}
}
\value{
a matrix with num_models columns and num_obs rows, with (log) weight
  for model m at observation i in entry [i, m]
}
\description{
Function to compute (log) weights of component models from the "predictions"
output by xgboost
}

