% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densitystackr-package.R
\docType{package}
\name{densitystackr}
\alias{densitystackr}
\alias{densitystackr-package}
\title{densitystackr - Stacking via xgboost}
\description{
An implementation of model stacking using xgboost.
}
\author{
Evan Ray <elray@umass.edu>
}

