% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking-estimation.R
\name{preds_to_matrix}
\alias{preds_to_matrix}
\title{Convert vector of predictions to matrix format
From line 56 at https://github.com/dmlc/xgboost/blob/ef4dcce7372dbc03b5066a614727f2a6dfcbd3bc/src/objective/multiclass_obj.cc,
it appears that preds is stored in column-major order with
observations in columns and classes/models in rows
i.e., preds[k * nclass + i] is prediction for model i at index k}
\usage{
preds_to_matrix(preds, num_models)
}
\arguments{
\item{preds}{vector of predictions as obtained from xgb.train}

\item{number}{of models}
}
\value{
preds in matrix form, with num_models columns and num_obs rows
}
\description{
Convert vector of predictions to matrix format
From line 56 at https://github.com/dmlc/xgboost/blob/ef4dcce7372dbc03b5066a614727f2a6dfcbd3bc/src/objective/multiclass_obj.cc,
it appears that preds is stored in column-major order with
observations in columns and classes/models in rows
i.e., preds[k * nclass + i] is prediction for model i at index k
}

